Adding viper flag allowing to disable journald logger
