+++ main.go
@@ -2,26 +2,9 @@ package main
 
 import (
 	"github.com/DSpeichert/netbootd/cmd"
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
-	"os"
 )
 
 //go:generate protoc proto/*.proto --go_out=. --go_opt=paths=source_relative --go-grpc_out=. --go-grpc_opt=paths=source_relative
-
-var (
-	journalLoggerEnabled bool
-)
-
-func init() {
-	// UNIX Time is faster and smaller than most timestamps
-	// If you set zerolog.TimeFieldFormat to an empty string,
-	// logs will write with UNIX time
-	zerolog.TimeFieldFormat = zerolog.TimeFormatUnix
-	if !journalLoggerEnabled {
-		log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr})
-	}
-}
 
 func main() {
 	cmd.Execute()
